add fn to step back through time
This function is not completely done yet. Right now the memory model
does not have support for stepping back because it's possible to step
back into an unitialized state.

The Diff struct also has not way to represent this. I'm reticent to make
the val field and option as this seems like a niche use case that would
pollute the rest of the code.