refactor Update struct
This should make it much easier to add more informantion to Updates. It
also allows us to add a warnings field, which is just a list of errors,
to Update, while keeping ProcessorUpdate clonable. Some other small
changes like dependency injecting the instruction for next_state, which
has been renamed to calculate_update