store reg return addr is stored in w/ stack frame
This allows us to detect if we called a function, stored the return
address in `reg`, and then tried to return to a return address stored in
a different register. This might be a bit of a nit, so maybe we'll only
emit this as a warning in the future.